+++ lib/meme/providers/my_meme_list_provider.dart
@@ -1,6 +1,6 @@
 import 'dart:typed_data';
 
-import 'package:hooks_riverpod/hooks_riverpod.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:myecl/auth/providers/openid_provider.dart';
 import 'package:myecl/meme/class/meme.dart';
 import 'package:myecl/meme/repositories/my_meme_repository.dart';
@@ -8,20 +8,26 @@ import 'package:myecl/tools/providers/list_notifier.dart';
 import 'package:myecl/tools/token_expire_wrapper.dart';
 
 class MyMemeListNotifier extends ListNotifier<Meme> {
-  final MyMemeRepository _myMemeRepository = MyMemeRepository();
+  final _myMemeRepository = MyMemeRepository();
   MyMemeListNotifier({required String token}) : super(const AsyncLoading()) {
     _myMemeRepository.setToken(token);
   }
 
   Future<AsyncValue<List<Meme>>> getMyMeme() async {
-    return await loadList(
-      () async => _myMemeRepository.getMyMeme(),
-    );
+    return await loadList(() async => _myMemeRepository.getMyMeme());
   }
 
   Future<bool> addMeme(Uint8List bytes) async {
     final success = await _myMemeRepository.addMeme(bytes);
     return success;
+  }
+
+  Future<Uint8List> getMemeImage(String id) async {
+    return await _myMemeRepository.getMemeImage(id);
+  }
+
+  Future<bool> deleteMeme(Meme meme) async {
+    return await _myMemeRepository.deleteMeme(meme.id);
   }
 }
